requirements.txt
altgraph==0.17
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
pyinstaller==4.5
pyinstaller-hooks-contrib==2021.2
python-telegram-bot==13.7
pytz==2021.1
selenium==3.141.0
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
